requirements.txt
# Dev environment
pip
autopep8

# Web framework
streamlit
requests
Pillow

# Mouse automation
pyautogui

# Video downloading
yt-dlp

# Browser automation
selenium
